Remove repeating #version and #extension directives
They can't appear in the middle of a shader, so we must remove the
repeated ones to avoid compilation errors if more than one source string
have been concatenated.
This is a quick-and-dirty fix: variations of directives like "#   version"
are not checked; the strings checked for might be part of a comment, which
we'll mistake for actual directives.